+++ unliftio/src/UnliftIO/Process.hs
@@ -1,7 +1,7 @@
 {-# LANGUAGE CPP #-}
 -- | Unlifted "System.Process".
 --
--- @since 0.2.4.0
+-- @since 0.2.5.0
 
 module UnliftIO.Process (
   -- * Running sub-processes
@@ -62,7 +62,7 @@ import qualified System.Process as P
 
 -- | Lifted 'P.createProcess'.
 --
--- @since 0.2.4.0
+-- @since 0.2.5.0
 {-# INLINE createProcess #-}
 createProcess ::
      MonadIO m
@@ -73,7 +73,7 @@ createProcess = liftIO . P.createProcess
 #if MIN_VERSION_process(1,2,1)
 -- | Lifted 'P.createProcess_'.
 --
--- @since 0.2.4.0
+-- @since 0.2.5.0
 {-# INLINE createProcess_ #-}
 createProcess_ ::
      MonadIO m
@@ -85,28 +85,28 @@ createProcess_ msg proc_ = liftIO (P.createProcess_ msg proc_)
 
 -- | Lifted 'P.callProcess'.
 --
--- @since 0.2.4.0
+-- @since 0.2.5.0
 {-# INLINE callProcess #-}
 callProcess :: MonadIO m => FilePath -> [String] -> m ()
 callProcess cmd args = liftIO (P.callProcess cmd args)
 
 -- | Lifted 'P.callCommand'.
 --
--- @since 0.2.4.0
+-- @since 0.2.5.0
 {-# INLINE callCommand #-}
 callCommand :: MonadIO m => String -> m ()
 callCommand = liftIO . P.callCommand
 
 -- | Lifted 'P.spawnProcess'.
 --
--- @since 0.2.4.0
+-- @since 0.2.5.0
 {-# INLINE spawnProcess #-}
 spawnProcess :: MonadIO m => FilePath -> [String] -> m ProcessHandle
 spawnProcess cmd args = liftIO (P.spawnProcess cmd args)
 
 -- | Lifted 'P.spawnCommand'.
 --
--- @since 0.2.4.0
+-- @since 0.2.5.0
 {-# INLINE spawnCommand #-}
 spawnCommand :: MonadIO m => String -> m ProcessHandle
 spawnCommand = liftIO . P.spawnCommand
@@ -114,7 +114,7 @@ spawnCommand = liftIO . P.spawnCommand
 #if MIN_VERSION_process(1,2,3)
 -- | Lifted 'P.readCreateProcess'.
 --
--- @since 0.2.4.0
+-- @since 0.2.5.0
 {-# INLINE readCreateProcess #-}
 readCreateProcess :: MonadIO m => CreateProcess -> String -> m String
 readCreateProcess cp input = liftIO (P.readCreateProcess cp input)
@@ -122,7 +122,7 @@ readCreateProcess cp input = liftIO (P.readCreateProcess cp input)
 
 -- | Lifted 'P.readProcess'.
 --
--- @since 0.2.4.0
+-- @since 0.2.5.0
 {-# INLINE readProcess #-}
 readProcess :: MonadIO m => FilePath -> [String] -> String -> m String
 readProcess cmd args input = liftIO (P.readProcess cmd args input)
@@ -130,7 +130,7 @@ readProcess cmd args input = liftIO (P.readProcess cmd args input)
 #if MIN_VERSION_process(1,2,3)
 -- | Lifted 'P.readCreateProcessWithExitCode'.
 --
--- @since 0.2.4.0
+-- @since 0.2.5.0
 {-# INLINE readCreateProcessWithExitCode #-}
 readCreateProcessWithExitCode ::
      MonadIO m => CreateProcess -> String -> m (ExitCode, String, String)
@@ -140,7 +140,7 @@ readCreateProcessWithExitCode cp input =
 
 -- | Lifted 'P.readProcessWithExitCode'.
 --
--- @since 0.2.4.0
+-- @since 0.2.5.0
 {-# INLINE readProcessWithExitCode #-}
 readProcessWithExitCode ::
      MonadIO m => FilePath -> [String] -> String -> m (ExitCode, String, String)
@@ -150,7 +150,7 @@ readProcessWithExitCode cmd args input =
 #if MIN_VERSION_process(1,4,3)
 -- | Unlifted 'P.withCreateProcess'.
 --
--- @since 0.2.4.0
+-- @since 0.2.5.0
 {-# INLINE withCreateProcess #-}
 withCreateProcess ::
      MonadUnliftIO m
@@ -168,28 +168,28 @@ withCreateProcess c action =
 
 -- | Lifted 'P.waitForProcess'.
 --
--- @since 0.2.4.0
+-- @since 0.2.5.0
 {-# INLINE waitForProcess #-}
 waitForProcess :: MonadIO m => ProcessHandle -> m ExitCode
 waitForProcess = liftIO . P.waitForProcess
 
 -- | Lifted 'P.getProcessExitCode'.
 --
--- @since 0.2.4.0
+-- @since 0.2.5.0
 {-# INLINE getProcessExitCode #-}
 getProcessExitCode :: MonadIO m => ProcessHandle -> m (Maybe ExitCode)
 getProcessExitCode = liftIO . P.getProcessExitCode
 
 -- | Lifted 'P.terminateProcess'.
 --
--- @since 0.2.4.0
+-- @since 0.2.5.0
 {-# INLINE terminateProcess #-}
 terminateProcess :: MonadIO m => ProcessHandle -> m ()
 terminateProcess = liftIO . P.terminateProcess
 
 -- | Lifted 'P.interruptProcessGroupOf'.
 --
--- @since 0.2.4.0
+-- @since 0.2.5.0
 {-# INLINE interruptProcessGroupOf #-}
 interruptProcessGroupOf :: MonadIO m => ProcessHandle -> m ()
 interruptProcessGroupOf = liftIO . P.interruptProcessGroupOf
@@ -197,7 +197,7 @@ interruptProcessGroupOf = liftIO . P.interruptProcessGroupOf
 #if MIN_VERSION_process(1,2,1)
 -- | Lifted 'P.createPipe'.
 --
--- @since 0.2.4.0
+-- @since 0.2.5.0
 {-# INLINE createPipe #-}
 createPipe :: MonadIO m => m (Handle, Handle)
 createPipe = liftIO P.createPipe
@@ -206,7 +206,7 @@ createPipe = liftIO P.createPipe
 #if MIN_VERSION_process(1,4,2)
 -- | Lifted 'P.createPipeFd'.
 --
--- @since 0.2.4.0
+-- @since 0.2.5.0
 {-# INLINE createPipeFd #-}
 createPipeFd :: MonadIO m => m (FD, FD)
 createPipeFd = liftIO P.createPipeFd