+++ unliftio/ChangeLog.md
@@ -1,8 +1,8 @@
-## 0.2.4.0
+## 0.2.5.0
 
 * Add `UnliftIO.Environment`/`UnliftIO.Foreign`/`UnliftIO.Process`
 
-## 0.2.3.0
+## 0.2.4.0
 
 * Use more generalized `withRunInIO` in `unliftio-core-0.1.1.0`
 * Add `getMonotonicTime` function